+++ cypress/integration/step1.spec.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
 
-import {MSG_ERROR_INVALID_NUMBER, MSG_ERROR_NO_NAMES} from "../../src/constants.mjs";
+import {MSG_ERROR_INVALID_NUMBER, MSG_ERROR_NO_NAMES} from "../../src/core/constants.mjs";
 import {initApp} from "../../src/app";
 
 describe('step1', () => {
