+++ src/core/constants.mjs
