+++ src/core/racing.mjs
@@ -1,4 +1,5 @@
 import {MSG_ERROR_INVALID_NUMBER, MSG_ERROR_NO_NAMES} from "./constants.mjs";
+import {isEmpty, validationTryCount} from "./validation.mjs";
 
 export const getRandomNumber = (min, max) =>
   Math.floor(Math.random() * (max + 1 - min) + min);
@@ -6,31 +7,30 @@ export const getRandomNumber = (min, max) =>
 export const [COMMAND_GO, COMMAND_STOP] = ["go", "stop"];
 export const getCommand = (number) => (number >= 4 ? COMMAND_GO : COMMAND_STOP);
 
-export function validationCarNames(str) {
-    const carsNames = str.split(",");
-
-    return !(carsNames.length === 0 || carsNames.some((name) => name.trim() === ""));
-}
-
 export function getCarsNames(str) {
-    if (validationCarNames(str)) {
+    if (isEmpty(str)) {
         return str.split(',');
     }
     throw new Error(MSG_ERROR_NO_NAMES);
 }
 
-export function validationTryCount(str) {
-    const count = parseInt(str, 10);
-
-    if (!count || isNaN(count)) {
-        return false;
-    }
-    return true;
-}
-
 export function getTryCount(str) {
     if (validationTryCount(str)) {
-        return parseInt(str);
+        return parseInt(str, 10);
     }
     throw new Error(MSG_ERROR_INVALID_NUMBER);
 }
+
+export const getWinners = (raceResult) => {
+    const goCounts = Object
+        .keys(raceResult)
+        .reduce((acc, key) => {
+            acc[key] = raceResult[key].filter(command => command === COMMAND_GO).length;
+            return acc;
+        }, {})
+    const winnerScore = Math.max(...Object.values(goCounts));
+
+    return Object.keys(goCounts).filter(
+        (key) => goCounts[key] === winnerScore
+    );
+};
