+++ src/core/templates.mjs
