--- src/module/constants.mjs
@@ -1,2 +1,0 @@
-export const MSG_ERROR_NO_NAMES = "자동차 이름을 입력해주세요.";
-export const MSG_ERROR_INVALID_NUMBER = "0이 아닌 숫자를 입력해주세요.";
